requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pypdf>=3.17.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
# OCR and image processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
# Table extraction
pdfplumber>=0.10.0
# LLM providers
openai>=1.0.0
google-generativeai>=0.3.0
# Additional utilities
python-dotenv>=1.0.0